refactor(direct-control): extract operation result types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves generic operation request result type ownership into src/play/operations/validate-request.ts and unit-operation, population-placement, and production postcondition public type ownership into their postcondition helper modules while keeping public facade type re-exports in index.ts.

Leaves production choice, diplomacy/narrative closeout, public procedure schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- unit-operation.test.ts population-placement.test.ts production-choice.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.play.operation-wrappers.test.ts game.play.population-placement.test.ts game.play.production.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle scans for operation result/postcondition definitions and facade executable imports

Local package/type relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
