+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -1182,6 +1182,19 @@ All future agent waves must be framed before delegation:
   semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
   matrix-row acceptance remain pending. This is local package/type relocation
   proof only, not runtime/live-game proof.
+- Direct-control operation result/postcondition types slice: completed as a
+  narrow public type ownership relocation. It moves generic operation request
+  result type ownership into `src/play/operations/validate-request.ts` and
+  unit-operation, population-placement, and production postcondition public
+  type ownership into their respective postcondition helper modules while
+  keeping public facade type re-exports in `index.ts`. This preserves public
+  type contracts and does not change operation source strings, wrappers,
+  validation, approval-first behavior, postcondition classification, or request
+  semantics. Production choice, diplomacy/narrative closeout, public procedure
+  schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI
+  semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
+  matrix-row acceptance remain pending. This is local package/type relocation
+  proof only, not runtime/live-game proof.
 - Direct-control diplomacy closeout source slice: completed as a narrow
   operation source relocation. It moves only the App UI diplomacy response
   closeout command builder and embedded closeout source into
